Seperate fetch functionality into fetchAPI file
